[+] Added Some Animations
