+++ src/sections/extra-pages/contact/ContactHeader.jsx
@@ -11,16 +11,13 @@ import AuthBackground from 'assets/images/auth/AuthBackground';
 
 export default function ContactHeader() {
   return (
-    <Box sx={{ position: 'relative', overflow: 'hidden', pt: 9, pb: 2 }}>
+    <Box sx={{ position: 'relative', overflow: 'hidden', pt: 9, pb: 0 }}>
       <AuthBackground />
       <Container maxWidth="lg" sx={{ px: { xs: 0, sm: 2 } }}>
         <Box sx={{ width: { xs: '100%', sm: 360, lg: 436 }, px: 2, py: 6, mx: 'auto' }}>
           <Stack sx={{ gap: 1 }}>
-            <Typography align="center" variant="h2">
+            <Typography align="center" variant="h3">
               Talk to our Expert
-            </Typography>
-            <Typography align="center" sx={{ color: 'text.secondary' }}>
-              Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
             </Typography>
           </Stack>
         </Box>
